# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blog API",
    "version": "1.0.0",
    "description": "Scalable Blog Backend API"
  },
  "servers": [
    {
      "url": "http://192.168.1.52:5000",
      "description": "Local network server"
    },
    {
      "url": "https://cinema-jackal-disclose.ngrok-free.dev/",
      "description": "ngrok tunnel"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "PostCreate": {
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "scheduled",
              "archived"
            ]
          },
          "media": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            }
          }
        }
      },
...
      "CommentCreate": {
        "type": "object",
        "required": [
          "postId",
          "content"
        ],
        "properties": {
          "postId": {
            "type": "string",
            "format": "uuid"
          },
          "content": {
            "type": "string"
          }
        }
      },
...
      "ReactionToggle": {
        "type": "object",
        "required": [
          "postId",
          "reactionType"
        ],
        "properties": {
          "postId": {
            "type": "string",
            "format": "uuid"
          },
          "reactionType": {
            "type": "string",
            "enum": [
              "like",
              "dislike"
            ]
          }
        }
      },
...
      "Profile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "avatar_url": {
            "type": "string",
            "format": "uri"
          },
          "followers_count": {
            "type": "integer"
          },
          "likes_count": {
            "type": "integer"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": [
              "object",
              "array",
              "string",
              "number",
              "boolean",
              "null"
            ]
          }
        }
      },
      "ProfileUpdate": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          }
        }
      },
...
      "AdminStats": {
        "type": "object",
        "properties": {
          "totalUsers": {
            "type": "integer"
          },
          "totalCreators": {
            "type": "integer"
          },
          "totalPublishedBlogs": {
            "type": "integer"
          },
          "blogsPublishedToday": {
            "type": "integer"
          },
          "totalComments": {
            "type": "integer"
          },
          "totalViews": {
            "type": "integer"
          },
          "totalLikes": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/posts": {
      "post": {
        "summary": "Create a new post",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PostCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all posts",
        "tags": [
          "Posts"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 5
            },
            "description": "Number of items per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search keyword for filtering by title or content"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Category for filtering posts"
          },
          {
            "in": "query",
            "name": "authorId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Author user ID to retrieve posts from a specific creator"
          },
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Alias for authorId"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "published",
                "scheduled",
                "archived"
              ]
            },
            "description": "Filter by post status (Defaults to 'published' for public queries; returns all status if authorId/userId is passed)"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "created_at",
                "published_at",
                "title"
              ],
              "default": "created_at"
            },
            "description": "Field to sort the posts by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ],
              "default": "DESC"
            },
            "description": "Sorting direction (ASC or DESC)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/comments": {
      "post": {
        "summary": "Create a new comment",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reactions": {
      "post": {
        "summary": "Toggle a reaction for a post",
        "tags": [
          "Reactions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactionToggle"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reaction toggled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/profiles": {
      "get": {
        "summary": "Get all public profiles",
        "tags": [
          "Profiles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter profiles by user role level (e.g. creator)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of profiles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Profile"
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update user profile details (fullName and bio)",
        "tags": [
          "Profiles"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/profile": {
      "get": {
        "summary": "Get authenticated user profile",
        "tags": [
          "General"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/analytics/admin": {
      "get": {
        "summary": "Get global platform statistics for administrators",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Admin platform summary stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStats"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admins only"
          }
        }
      }
    },
...
  }
}